Move services list to separate file, add services status on startup
