+++ server-side/app.js
@@ -5,6 +5,7 @@ const bodyParser = require("body-parser");
 const cors = require("cors");
 
 const NCIRouter = require("./routes/NCIRouter");
+const ServicesList = require("./servicesList");
 
 const hostname = "localhost";
 const port = process.env.NCI_BACKEND_PORT || 8000;
@@ -26,4 +27,12 @@ app.use((req, res, next) => {
 const server = http.createServer(app);
 server.listen(port, hostname, () => {
   console.log(`Backend is running at http://${hostname}:${port}`);
+  const services = ServicesList.getServicesList();
+  if(services[0].name == "PlaceholderService"){
+    console.log('No services are enabled! Only placeholder results will be shown!');
+  }
+  else
+  {
+    console.log(`Enabled services [${services.length}/${ServicesList.getAllServices().length}]:`, services.map(s => s.name).join(', '));
+  }
 });
