+++ server-side/routes/NCIRouter.js
@@ -1,16 +1,15 @@
 const express = require("express");
 const bodyParser = require("body-parser");
+const ServicesList = require("../servicesList");
+
 const NCIRouter = express.Router();
 NCIRouter.use(bodyParser.json());
-const PlaceholderService = require('../services/placeholderService')
-const PixabayService = require('../services/pixabayService')
 
-const imageCount = 50;
 
 NCIRouter.route("/").get((req, res, next) => {
   const query = req.query.img;
   const promises = [];
-  for(const service of getServicesList())
+  for(const service of ServicesList.getServicesList())
   {
     promises.push(new service().request(query));
   }
@@ -22,7 +21,7 @@ NCIRouter.route("/").get((req, res, next) => {
     .then(result => {
       for (let i = 0; i < result.length; i++) {
         const serviceResult = result[i];
-        const serviceName = getServicesList()[i].name;
+        const serviceName = ServicesList.getServicesList()[i].name;
         if(serviceResult instanceof Error) {
           console.log(`Service failed - ${serviceName}: `, serviceResult)
           unsuccessfulServices.push(serviceName);
@@ -51,30 +50,5 @@ NCIRouter.route("/").get((req, res, next) => {
     });
 });
 
-function getServicesList()
-{
-  const services = [ PixabayService ]
-  const enabledServices = [];
-  for(const service of services) {
-    if(new service().isEnabled()) {
-      enabledServices.push(service);
-    }
-  }
-  if(enabledServices.length > 0) {
-    return enabledServices
-  }
-  else {
-    return [ PlaceholderService ];
-  }
-}
-
-function shuffleArray(array) {
-  for (var i = array.length - 1; i > 0; i--) {
-    var j = Math.floor(Math.random() * (i + 1));
-    var temp = array[i];
-    array[i] = array[j];
-    array[j] = temp;
-  }
-}
 
 module.exports = NCIRouter;
